src/FlowSE/requirement.txt
# torch==2.2.0
# torchaudio==2.2.0
# triton==2.2.0
# numpy==1.26.4

absl-py==2.1.0
accelerate==1.2.1
einops==0.8.0
einx==0.3.0
ema-pytorch==0.7.7
safetensors==0.4.5
x-transformers==1.43.2
torch-complex==0.4.4
torchdiffeq==0.2.5
transformers==4.38.2
peft==0.10.0

librosa==0.10.1
soundfile==0.12.1
vocos==0.1.0
pydub==0.25.1
audioread==3.0.1
julius==0.2.7
resampy==0.4.3
soxr==0.3.7
jiwer
s3prl
openai-whisper

huggingface-hub==0.32.0
hf_transfer
sentencepiece==0.2.0
jieba==0.42.1
pypinyin==0.53.0
jamo==0.4.1
jaconv==0.4.0

PyYAML==6.0.1
tqdm==4.66.4
scipy==1.14.1
pandas==2.1.3
matplotlib==3.9.1
seaborn==0.13.2
requests==2.31.0
httpx==0.28.1
loguru==0.7.3
tensorboard

# onnxruntime-gpu==1.12.0
# wespeaker (请通过 git+https://github.com/wenet-e2e/wespeaker.git 安装)